Addressed PR comments, added color property to mixins.
