+++ src/components/MainSubHeader/MainSubHeader.scss
@@ -1,0 +1,60 @@
+@use "./../../styles/partials/global" as *;
+@use "./../../styles/partials/variables" as *;
+@use "./../../styles/partials/mixins" as *;
+
+.header {
+    &__container {
+        display: flex;
+        flex-direction: column;
+        padding-top: 32px;
+        margin-left: 24px;
+        margin-right: 24px;
+
+        @include tablet {
+            flex-direction: row;
+            justify-content: space-between;
+            align-items: center;
+            padding: 32px;
+            margin: 0;
+        }
+
+        @include desktop {
+            padding-left: 40px;
+            padding-right: 40px;
+        }
+
+        & > div {
+            display: flex;
+            flex-direction: column;
+
+            @include tablet {
+                flex-direction: row;
+                justify-content: flex-start;
+            }
+        }
+    }
+
+    &__title {
+        @include page-header;
+    }
+
+    &__search-bar {
+        margin-top: 16px;
+        margin-bottom: 16px;
+        display: flex;
+        justify-content: center;
+        border: 1px solid #bdc5d5;
+        border-radius: 20px;
+        padding: 8px;
+        padding-left: 16px;
+        background-image: url("../../assets/icons/search-24px.svg");
+        background-repeat: no-repeat;
+        background-position: 95%;
+
+        @include tablet {
+            margin-top: 0;
+            margin-right: 16px;
+            height: 36px;
+        }
+    }
+}
