--- src/components/WarehouseList/WarehouseList.scss
@@ -1,148 +1,0 @@
-// awaiting SCSS partials
-
-// temp Media Breakpoints
-$tablet-width: 768px;
-$desktop-width: 1280px;
-
-@mixin tablet {
-  @media (min-width: $tablet-width) {
-    @content;
-  }
-}
-
-@mixin desktop {
-  @media (min-width: $desktop-width) {
-    @content;
-  }
-}
-
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.warehouse-list {
-  margin: 0 16px;
-  box-shadow: 0px 2px 5px rgba(19, 24, 44, 0.1);
-  border-radius: 3px;
-
-  @include tablet {
-    margin-left: 32px;
-    margin-right: 32px;
-  }
-
-  @include desktop {
-    width: 1020px;
-    margin: auto;
-  }
-
-  &__header {
-    display: flex;
-    flex-direction: column;
-    padding-top: 32px;
-    margin-left: 24px;
-    margin-right: 24px;
-
-    @include tablet {
-      flex-direction: row;
-      justify-content: space-between;
-      padding: 32px;
-      margin: 0;
-    }
-
-    @include desktop {
-      padding-left: 40px;
-      padding-right: 40px;
-    }
-
-    & > div {
-      display: flex;
-      flex-direction: column;
-
-      @include tablet {
-        flex-direction: row;
-        justify-content: flex-start;
-      }
-    }
-  }
-
-  &__search-bar {
-    margin-top: 16px;
-    display: flex;
-    justify-content: center;
-    border: 1px solid #bdc5d5;
-    border-radius: 20px;
-    padding: 8px;
-    padding-left: 16px;
-    background-image: url("../../assets/icons/search-24px.svg");
-    background-repeat: no-repeat;
-    background-position: 95%;
-
-    @include tablet {
-      margin-top: 0;
-      margin-right: 16px;
-      height: 36px;
-    }
-  }
-
-  &__add-btn {
-    margin-top: 16px;
-    // background instock-indigo
-    // font white
-    border-radius: 20px;
-    height: 36px;
-
-    @include tablet {
-      margin-top: 0;
-      padding: 8px 16px;
-      height: 38px;
-    }
-  }
-
-  &__columns {
-    display: none;
-    width: 100%;
-
-    @include tablet {
-      display: grid;
-      grid-template-columns: 1fr 1fr 1fr 1.3fr 1fr;
-      background-color: #f7f8f9;
-      padding-top: 11px;
-      padding-bottom: 11px;
-      padding-left: 32px;
-      padding-right: 32px;
-    }
-
-    @include desktop {
-      grid-template-columns: 1fr 1.9fr 1fr 1fr 1fr;
-    }
-
-    h3 {
-      font-size: 12px;
-    }
-
-    img {
-      height: 12px;
-    }
-
-    & > div {
-      display: flex;
-      align-items: center;
-
-      &:last-child {
-        justify-content: flex-end;
-      }
-    }
-  }
-
-  &__list {
-    margin-top: 24px;
-    margin-left: 24px;
-    margin-right: 24px;
-
-    @include tablet {
-      margin-top: 0;
-    }
-  }
-}
