+++ src/components/WarehousesList/WarehousesList.scss
@@ -1,0 +1,76 @@
+@use "./../../styles/partials/global" as *;
+@use "./../../styles/partials/variables" as *;
+@use "./../../styles/partials/mixins" as *;
+
+.warehouse-list {
+    margin: 0 16px;
+    box-shadow: 0px 2px 5px rgba(19, 24, 44, 0.1);
+    border-radius: 3px;
+    background-color: $white;
+
+    @include tablet {
+        margin-left: 32px;
+        margin-right: 32px;
+    }
+
+    @include desktop {
+        width: 1020px;
+        margin: auto;
+    }
+
+    & a {
+        @include labels;
+        text-decoration: none;
+        color: $instock-indigo;
+
+        &:visited {
+            color: $instock-indigo;
+        }
+    }
+
+    &__columns {
+        display: none;
+        width: 100%;
+
+        @include tablet {
+            display: grid;
+            grid-template-columns: 1fr 1fr 1fr 1.3fr 1fr;
+            background-color: $light-grey;
+            padding-top: 11px;
+            padding-bottom: 11px;
+            padding-left: 32px;
+            padding-right: 32px;
+        }
+
+        @include desktop {
+            grid-template-columns: 1fr 1.9fr 1fr 1fr 1fr;
+        }
+
+        h3 {
+            @include table-header;
+        }
+
+        img {
+            height: 12px;
+        }
+
+        & > div {
+            display: flex;
+            align-items: center;
+
+            &:last-child {
+                justify-content: flex-end;
+            }
+        }
+    }
+
+    &__list {
+        margin-top: 24px;
+        margin-left: -16px;
+        margin-right: 24px;
+
+        @include tablet {
+            margin-top: 0;
+        }
+    }
+}
