+++ src/components/WarehousesListColumns/WarehousesListColumns.scss
@@ -1,0 +1,39 @@
+@use "./../../styles/partials/global" as *;
+@use "./../../styles/partials/variables" as *;
+@use "./../../styles/partials/mixins" as *;
+
+.warehouse-list__columns {
+    display: none;
+    width: 100%;
+
+    @include tablet {
+        display: grid;
+        grid-template-columns: 1fr 1fr 1fr 1.3fr 1fr;
+        background-color: $light-grey;
+        padding-top: 11px;
+        padding-bottom: 11px;
+        padding-left: 32px;
+        padding-right: 32px;
+    }
+
+    @include desktop {
+        grid-template-columns: 1fr 1.9fr 1fr 1fr 1fr;
+    }
+
+    h3 {
+        @include table-header;
+    }
+
+    img {
+        height: 12px;
+    }
+
+    & > div {
+        display: flex;
+        align-items: center;
+
+        &:last-child {
+            justify-content: flex-end;
+        }
+    }
+}
