+++ src/components/index.js
@@ -1,4 +1,5 @@
 export { default as WarehousesList } from "./WarehousesList/WarehousesList";
+export { default as WarehousesListItem } from "./WarehousesListItem/WarehousesListItem";
 export { default as AddNewWarehouse } from "./AddNewWarehouse/AddNewWarehouse";
 export { default as SelectedWarehouse } from "./SelectedWarehouse/SelectedWarehouse";
 export { default as EditWarehouse } from "./EditWarehouse/EditWarehouse";
@@ -9,5 +10,12 @@ export { default as AddNewItem } from "./AddNewItem/AddNewItem";
 export { default as SelectedItem } from "./SelectedItem/SelectedItem";
 export { default as EditItem } from "./EditItem/EditItem";
 export { default as DeleteItem } from "./DeleteItem/DeleteItem";
+
+export { default as BlueButton } from "./Buttons/BlueButton/BlueButton";
+export { default as RedButton } from "./Buttons/RedButton/RedButton";
+export { default as WhiteButton } from "./Buttons/WhiteButton/WhiteButton";
+
+export { default as MainSubHeader } from "./MainSubHeader/MainSubHeader";
+export { default as WarehousesListColumns } from "./WarehousesListColumns/WarehousesListColumns";
 
 export { default as Footer } from "./Footer/Footer";
