+++ src/styles/partials/_mixins.scss
@@ -25,6 +25,7 @@
     font-family: TitilliumWeb-SemiBold;
     font-size: 28px;
     line-height: 36px;
+    color: $instock-black;
 }
 
 @mixin subheader {
@@ -35,6 +36,7 @@
     font-family: TitilliumWeb-SemiBold;
     font-size: 20px;
     line-height: 28px;
+    color: $instock-black;
 }
 
 @mixin labels {
@@ -55,6 +57,7 @@
     font-family: TitilliumWeb-SemiBold;
     font-size: 11px;
     line-height: 16px;
+    color: $slate;
 }
 
 @mixin body-large {
@@ -65,6 +68,7 @@
     font-family: TitilliumWeb-Regular;
     font-size: 15px;
     line-height: 26px;
+    color: $instock-black;
 }
 
 @mixin body-medium {
@@ -75,6 +79,7 @@
     font-family: TitilliumWeb-Regular;
     font-size: 13px;
     line-height: 20px;
+    color: $instock-black;
 }
 
 @mixin body-small {
@@ -85,6 +90,7 @@
     font-family: TitilliumWeb-Regular;
     font-size: 11px;
     line-height: 16px;
+    color: $instock-black;
 }
 
 // Note: if you set this on your outer most container, it will do the margins automatically for all 3 media. Can remove if prefer not to use this.
